File::Spec::Unix::catfile() should preserve a leading ./
This makes File::Spec::Unix::catfile('./a') and catfile('.', 'a') both
return './a' (the former used to return 'a'), while catfile('a') still
returns 'a'. This way, catfile() is more consistent, in that it really
concatenates its arguments. Repeated slashes and ./ inside the path
are still stripped.

Another consistent solution would have been to apply canonpath() before
returning the concatenated path, but this would have removed a leading
./ and it is sometimes necessary to force it (e.g. when you want to pass
the path to system()).

As a result of this change, catdir() and canonpath() are no longer
called by catfile(),

This fixes RT #113898.