don't call assign operator overloads on the rhs [perl #113834]
This fixes f041cf0, which made assignment operators not fall back to
their base operator when trying to swap the order of the arguments. In
the process of fixing that, it also made it try the overloaded version
of the assignment operator itself as well, which is wrong, as that
should only happen on the left side.

This still has problems - see <20120629185035.GA30375@tozt.net> on the
mailing list.