[perl #113012] String negation under ‘use integer’
This makes the negation operator under the integer pragma (i_int) use
the same logic for determining whether to do string negation as the
regular negation operator.

Before, this did not happen at all under the integer pragma, except
for barewords, resulting in strange inconsistencies:

$ perl -le 'use integer; print -foo'
-foo
$ perl -le 'use integer; print -"foo"'
0

The code for string negation is now in a static routine in pp.c and is
used by both types of negation.