ensure flags are always correct in sv_gets [perl #113906]
If a readline operation was interrupted by a signal handler, and that
signal handler overwrote the target of the readline assignment, the
string in the SvPV slot would be updated properly, but the sv flags
would all still correspond to the sv that was assigned in the signal
handler, rather than what should be expected by the readline operation.
For instance, if the signal handler did $ret = 100, $ret would still
have a string written to it, but it would be SvIOK and not SvPOK. This
fixes up the flags after the actual blocking read to make sure they are
still sane.