perlfunc: clarify docs for 'package' [perl #113974]
This one word change clarifies that 'package' applies to
'lexically-scoped' variables rather than 'lexical' variables, which
people may misunderstand to mean only my/state declarations and thus be
confused by the nearby statements about it applying to 'our' as well.

No perldelta note was added as the change was trivial.