Perl_ck_select() doesn't need fold_constants or op_integerize for PP_SSELECT.
Commit 985b9e549fd48480 refactored S_fold_constants() to split out the code
unrelated to constant folding into S_op_integerize() and S_op_std_init().
However, it simply replaced all calls to the former S_fold_constants() with
a chain of calls to all 3. For the case of PP_SSELECT, the calls to
S_op_integerize() and S_fold_constants() are redundant, as 4 argument select
does not have an integer form, and cannot be constant folded.