Tighten the use of local variables and LINKLIST() in S_fold_constants().
'type' should be *const* instead of volatile! (And of type OPCODE)
'curop' is only needed for the for loop that follows the op_next pointers to
validate that all child ops are of acceptable types.

Also refactor that loop to only call LINKLIST() once, instead of once per
loop iteration. The first call to LINKLIST() will set all the op_next
pointers correctly, so they can be dereferenced directly - no need for the
extra conditional code to call linklist() if one is NULL, because none will
be. The final call to LINKLIST() just prior to setting PL_op is also
redundant so remove it too.