Turn an assert() in S_fold_constants() into a panic.
Hence the failure diagnostic is now clearer about which precondition the
caller failed to meet, and the failure can be trapped by eval. (Which is
reasonable in this case - the interpreter is not in an unrecoverable state.)

Move the check earlier, so that the croak() is before the changes to PL_op,
else the panic itself will panic (which is somewhat counterproductive).