perlhacktips: PERL_DEBUG_READONLY_OPS update
I don’t know when this changed, but I *can* build F<perl> with
PERL_DEBUG_READONLY_OPS, and I could before I rewrote it to use the
new slab allocator.  So that particular note in perlhacktips may have
been wrong for quite some time.  I assume it had to do with redefini-
tion sub warnings that try to set the line number of the current cop
(which is still a problem).