[perl #113576] Restore err msg for &{+undef}
This ‘Can't use an undefined value as a subroutine reference’ message
used to occur for &{+undef}, until commit 15ff848f77796 in 1997.

Then it started happening only for magic variables.

Non-magical variables would treat it the same as "", possibly dying
with ‘Undefined subroutine main:: called’, and possibly calling the
subroutine with that name.

Magical variables (&{$tied_undef}) started behaving the same way
(treating undef as "") with commit 7c75014e, in 2010.

This commit restores that message.