perly.y: Remove use of latefree from package foo {}
It is not necessary for the op containing the sv containing the name
of the package to last until the end of the block.  Perl_package can
free the op immediately, as the only information it needs from it it
copies to the sv in PL_curstname.  The version number can be treated
the same way.