fold_constants() never folds OP_RV2GV, and hence never returns a newGVOP.
The code in fold_constants() that returns a newGVOP if the folded op is of
type OP_RV2GV has been present since 5.000 alpha 2 (or earlier), yet rv2gv
has never been flagged as eligible for constant folding. Hence this has been
dead code since it was added, and is safe to remove.