op.c:op_free: Rmv dead code; simplify cop_free logic
This reverts c53f1caa and cc93af5f.

See the thread starting at
http://www.nntp.perl.org/group/perl.perl5.porters/2008/04/msg135885.html

Basically, change c53f1caa made a change, but then cc93af5f undid it,
but differently.  This resulted in dead code left by c53f1caa (type is
unused after the assignment).  And in the end the code behaved exactly
the same way, so the original problem was not fixed.  I suspect this
was a B::C bug.