op.c: op_clear is tempting fate
This if() statement can be reached by op types to which the OpTRANS*
flags to not apply.  They happen at present not to use any flags that
conflict with these (except when OPf_KIDS is set, in which case this
code is not reached).   But we should make sure, via an assertion,
that new flags added to goto or last do not conflict with trans utf8
flags, and that trans utf8 flags (1 and 2), if renumbered, do not con-
flict with goto/last utf8 flags (128).