Restore microperl, which has been unable to build since January.
microperl was broken by commit 82ad65bb0613be64 on 2012-01-16, which used
IN_LOCALE_COMPILETIME for the first time in the C code. Unlike
IN_LOCALE_RUNTIME, it had no fallback definition for microperl.

Commit f90a9a0230170cc0 on 2012-01-28 added the first use of Strtol(),
which means that microperl now needs to assume that the system has strtol().
(Which is not unreasonable, as it's part of C89).