Make undef &foo remove call checkers
The fact that the call checker is stored in magic is an implementation
detail.  cv_undef does not free magic, so the call checker lives on.

If we were to move the parameter prototype into magic internally, we
would not want undef to stop clearing it.  To me, the current situa-
tion with call checkers is similar.